Fix: update component structure
