+++ src/utils.js
@@ -98,8 +98,3 @@ export const openLink = link => {
     window.location.replace(link)
   }
 }
-
-export const spinalCase = str => {
-  if (!str) return
-  return str.replace(/(?!^)([\s])/g, ' $1').replace(/[_\s]+(?=[a-zA-Z])/g, '-').toLowerCase()
-}
