+++ src/v1.1/components/FAQSection/FAQItem/faqItem.sass
@@ -1,19 +1,10 @@
-@import '../../styles/variables'
-@import '../../styles/mixins'
-@import '../../styles/colors'
-@import '../../styles/fonts'
+@import '~styles/variables'
+@import '~styles/mixins'
+@import '~styles/colors'
+@import '~styles/fonts'
 
 $open-hspace: 20px
 $open-hspace-xs: 30px
-
-// Default
-.faqsSection
-  padding-top: 100px
-  &__wrap
-    +grid()
-  &__title
-    grid-column: 2/span 8
-    margin-bottom: 32px
 
 .faqItem
   $c: &
@@ -79,23 +70,12 @@ $open-hspace-xs: 30px
 
 // XSmall
 +render-to($media-xs)
-  .faqsSection
-    padding-top: 70px
-    &__wrap
-      +grid-small()
-    &__title
-      grid-column: 1/span 10
-      padding-left: $hspace-s
   .faqItem
     grid-column: 1/span 10
     padding-left: $hspace-s
 
 // XXSmall
 +render-to($media-xxs)
-  .faqsSection
-    padding-top: 50px
-    &__title
-      padding-left: $hspace-xs
   .faqItem
     $c: &
     padding-left: $hspace-xs
