+++ src/v1.1/components/FAQSection/FAQItem/index.js
@@ -1,0 +1,2 @@
+import FAQItem from './FAQItem'
+export default FAQItem
