+++ src/v1.1/components/FAQSection/index.js
@@ -1,0 +1,2 @@
+import FAQSection from './FAQSection'
+export default FAQSection
