+++ src/v1.1/pages/FAQs/index.js
@@ -1,0 +1,2 @@
+import FAQs from './FAQs'
+export default FAQs
